+++ tests/testthat/test-write.R
@@ -35,6 +35,24 @@ test_that("writeH5AD works as expected", {
     expect_identical(col_data, colData(sce))
 })
 
+test_that("writeH5AD works as expected with sparse matrices", {
+    mat <- assay(sce)
+    counts(sce) <- as(mat, "dgCMatrix")
+    logcounts(sce) <- counts(sce) * 10
+    assay(sce, "random") <- mat # throwing in a dense matrix in a mixture.
+
+    temp <- tempfile(fileext = '.h5ad')
+    writeH5AD(sce, temp)
+    expect_true(file.exists(temp))
+
+    # Reading it back out again. Hopefully we didn't lose anything important.
+    out <- readH5AD(temp)
+    
+    expect_identical(counts(sce), assay(out, "X"))
+    expect_identical(logcounts(sce), logcounts(out))
+    expect_identical(assay(sce, "random"), assay(out, "random"))
+})
+
 test_that("writeH5AD works with assay skipping", {
     temp <- tempfile(fileext = '.h5ad')
     writeH5AD(sce, temp, skip_assays = TRUE)